revert package move for FrequencyRelations
