+++ docs/examples/units-extension/qual/Hz.java
@@ -8,7 +8,6 @@ import java.lang.annotation.Target;
 import org.checkerframework.checker.units.qual.Prefix;
 import org.checkerframework.checker.units.qual.UnitsRelations;
 import org.checkerframework.framework.qual.SubtypeOf;
-import relations.FrequencyRelations;
 
 /** Hertz (Hz), a unit of frequency. */
 @Documented
